Add files via upload
This file contains the viruses but requires a fix to the drop blocks methods so that the viruses do not drop as well. Change the colours of the viruses by one hex digit and check for colour when doing the drop block method, also ensure they still clear when 4 in a row are on the board. Create new method for set_colour_virus_green, set_colour_virus_blue, and set_colour_virus_red which are the same as the ones we have now but of course with the new colours for the viruses.